requirements.txt
# Packages used for this project

attrs==23.2.0
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
fake-useragent==1.5.1
Faker==26.0.0
filelock==3.15.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
packaging==24.1
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-box==7.2.0
python-dateutil==2.9.0.post0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
scrapy-fake-useragent==1.4.4
service-identity==24.1.0
setuptools==70.3.0
six==1.16.0
tldextract==5.1.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
zope.interface==6.4.post2